Add files via upload
This is the 1st visual as bar graph noting that delays are accumulated in heavier fashion in the later part of the year and the very beginning of the year.  It is proposed that if a customer inputs a flight for the summer months June to August there will be a high chance of delay.